+++ tests/Fixture/DelegatedClass.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Northwoods\Container\Fixture;
-
-class DelegatedClass
-{
-}