+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0]
+
+### Added
+
+- `Zend\Config` and `Zend\ContainerFactory` for improved Zend Service Manager compatibility
+
+### Removed
+
+- `Config\ServiceConfig` has been replaced with `Zend\Config`
+
 ## [2.1.0]
 
 ### Added